split out current loki actions into separate files; still need to remove unneeded code
