+++ src/loki/route_action.cc
@@ -1,0 +1,139 @@
+#include "loki/service.h"
+#include "loki/search.h"
+
+#include <boost/property_tree/info_parser.hpp>
+
+#include <valhalla/baldr/json.h>
+#include <valhalla/midgard/distanceapproximator.h>
+#include <valhalla/midgard/logging.h>
+
+using namespace prime_server;
+using namespace valhalla::baldr;
+
+namespace {
+  enum ACTION_TYPE {ROUTE, VIAROUTE, LOCATE};
+  const std::unordered_map<std::string, ACTION_TYPE> ACTION{
+    {"/route", ROUTE},
+    {"/viaroute", VIAROUTE},
+    {"/locate", LOCATE}
+  };
+  const headers_t::value_type CORS{"Access-Control-Allow-Origin", "*"};
+  const headers_t::value_type JSON_MIME{"Content-type", "application/json;charset=utf-8"};
+  const headers_t::value_type JS_MIME{"Content-type", "application/javascript;charset=utf-8"};
+
+  //TODO: move json header to baldr
+  //TODO: make objects serialize themselves
+
+  json::ArrayPtr serialize_edges(const PathLocation& location, GraphReader& reader, bool verbose) {
+    auto array = json::array({});
+    std::unordered_multimap<uint64_t, PointLL> ids;
+    for(const auto& edge : location.edges()) {
+      try {
+        //get the osm way id
+        auto tile = reader.GetGraphTile(edge.id);
+        auto* directed_edge = tile->directededge(edge.id);
+        auto edge_info = tile->edgeinfo(directed_edge->edgeinfo_offset());
+        //check if we did this one before
+        auto range = ids.equal_range(edge_info->wayid());
+        bool duplicate = false;
+        for(auto id = range.first; id != range.second; ++id) {
+          if(id->second == location.vertex()) {
+            duplicate = true;
+            break;
+          }
+        }
+        //only serialize it if we didnt do it before
+        if(!duplicate) {
+          ids.emplace(edge_info->wayid(), location.vertex());
+          //they want MOAR!
+          if(verbose) {
+            array->emplace_back(
+              json::map({
+                {"way_id", edge_info->wayid()},
+                {"correlated_lat", json::fp_t{location.vertex().lat(), 6}},
+                {"correlated_lon", json::fp_t{location.vertex().lng(), 6}}
+              })
+            );
+          }//spare the details
+          else {
+            array->emplace_back(
+              json::map({
+                {"way_id", edge_info->wayid()},
+                {"correlated_lat", json::fp_t{location.vertex().lat(), 6}},
+                {"correlated_lon", json::fp_t{location.vertex().lng(), 6}}
+              })
+            );
+          }
+        }
+      }
+      catch(...) {
+        //this really shouldnt ever get hit
+        LOG_WARN("Expected edge not found in graph but found by loki::search!");
+      }
+    }
+    return array;
+  }
+
+  json::MapPtr serialize(const PathLocation& location, GraphReader& reader, bool verbose) {
+    return json::map({
+      {"ways", serialize_edges(location, reader, verbose)},
+      {"input_lat", json::fp_t{location.latlng_.lat(), 6}},
+      {"input_lon", json::fp_t{location.latlng_.lng(), 6}}
+    });
+  }
+
+  json::MapPtr serialize(const PointLL& ll, const std::string& reason) {
+    return json::map({
+      {"ways", static_cast<std::nullptr_t>(nullptr)},
+      {"input_lat", json::fp_t{ll.lat(), 6}},
+      {"input_lon", json::fp_t{ll.lng(), 6}},
+      {"reason", reason}
+    });
+  }
+}
+
+namespace valhalla {
+  namespace loki {
+
+    worker_t::result_t loki_worker_t::route(const ACTION_TYPE& action, boost::property_tree::ptree& request, http_request_t::info_t& request_info) {
+      //see if any locations pairs are unreachable or too far apart
+      auto lowest_level = reader.GetTileHierarchy().levels().rbegin();
+      auto max_distance = config.get<float>("service_limits.max_distance." + request.get<std::string>("costing"));
+      for(auto location = ++locations.cbegin(); location != locations.cend(); ++location) {
+
+        //check connectivity
+        uint32_t a_id = lowest_level->second.tiles.TileId(std::prev(location)->latlng_);
+        uint32_t b_id = lowest_level->second.tiles.TileId(location->latlng_);
+        if(!reader.AreConnected({a_id, lowest_level->first, 0}, {b_id, lowest_level->first, 0}))
+          throw std::runtime_error("Locations are in unconnected regions. Go check/edit the map at osm.org");
+
+        //check if distance between latlngs exceed max distance limit for each mode of travel
+        auto path_distance = std::sqrt(midgard::DistanceApproximator::DistanceSquared(std::prev(location)->latlng_, location->latlng_));
+        if (path_distance > max_distance)
+          throw std::runtime_error("Path distance exceeds the max distance limit.");
+
+        LOG_INFO("location_distance::" + std::to_string(path_distance));
+      }
+
+      //correlate the various locations to the underlying graph
+      for(size_t i = 0; i < locations.size(); ++i) {
+        auto correlated = loki::Search(locations[i], reader, costing_filter);
+        request.put_child("correlated_" + std::to_string(i), correlated.ToPtree(i));
+      }
+
+      //let tyr know if its valhalla or osrm format
+      if(action == loki_worker_t::VIAROUTE)
+        request.put("osrm", "compatibility");
+      std::stringstream stream;
+      boost::property_tree::write_info(stream, request);
+
+      //ok send on the request with correlated origin and destination filled out
+      //using the boost ptree info format
+      //TODO: make a protobuf request object and pass that along, can be come
+      //part of thors path proto object and then get copied into odins trip object
+      worker_t::result_t result{true};
+      result.messages.emplace_back(stream.str());
+      return result;
+    }
+  }
+}
